Protecting Resources with Authorization

+++ backend/routes/posts.js
@@ -61,9 +61,12 @@ router.put("/:id", checkAuth, multer({storage: storage}).single("image"), (req, 
     content: req.body.content,
     imagePath: imagePath
   });
-  Post.updateOne({ _id: req.params.id }, post).then((result) => {
-    console.log(result);
-    res.status(200).json({ message: "Update successful!" });
+  Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then((result) => {
+    if(result.modifiedCount > 0) {
+      res.status(200).json({ message: "Update successful!" });
+    } else {
+      res.status(401).json({ message: "Not  authorized!"});
+    }
   });
 });
 
@@ -100,10 +103,12 @@ router.get("/:id", (req, res, next) => {
 });
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  console.log(req.params.id);
-  Post.deleteOne({ _id: req.params.id }).then((result) => {
-    console.log(result);
-    res.status(200).json({ message: "Post deleted!" });
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
+    if(result.deletedCount  > 0) {
+      res.status(200).json({ message: "Deletion deleted!" });
+    } else {
+      res.status(401).json({ message: "Not  authorized!"});
+    }
   });
 });
